pyproject.toml
[tool.poetry]
name = "fullstack"
version = "0.1.0"
description = ""
authors = ["tetradui <kutmanalown@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=5.0.6,<5.1.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.3.0"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
redis = "^5.0.4"
celery = "^5.4.0"
setuptools = "^69.5.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
